requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
absl-py=0.13.0=pypi_0
appnope=0.1.2=py39hecd8cb5_1001
argon2-cffi=20.1.0=py39h9ed2024_1
astunparse=1.6.3=pypi_0
async_generator=1.10=pyhd3eb1b0_0
attrs=21.2.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
bleach=3.3.1=pyhd3eb1b0_0
bzip2=1.0.8=h0d85af4_4
c-ares=1.17.1=h0d85af4_1
ca-certificates=2021.7.5=hecd8cb5_1
cachetools=4.2.2=pypi_0
cairo=1.16.0=he43a7df_1008
certifi=2021.5.30=py39hecd8cb5_0
cffi=1.14.6=py39h2125817_0
charset-normalizer=2.0.3=pypi_0
click=8.0.1=pypi_0
decorator=5.0.9=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
dlib=19.22.0=pypi_0
entrypoints=0.3=py39hecd8cb5_0
face-recognition=1.3.0=pypi_0
face-recognition-models=0.3.0=pypi_0
ffmpeg=4.3.1=hb787a4c_2
flatbuffers=1.12=pypi_0
fontconfig=2.13.1=h10f422b_1005
freetype=2.10.4=h4cff582_1
gast=0.4.0=pypi_0
gettext=0.19.8.1=h7937167_1005
gmp=6.2.1=h2e338ed_0
gnutls=3.6.13=h756fd2b_1
google-auth=1.33.0=pypi_0
google-auth-oauthlib=0.4.4=pypi_0
google-pasta=0.2.0=pypi_0
graphite2=1.3.13=h2e338ed_1001
grpcio=1.34.1=pypi_0
h5py=3.1.0=pypi_0
harfbuzz=2.8.2=h159f659_0
hdf5=1.10.6=nompi_hc5d9132_1114
icu=68.1=h74dc148_0
idna=3.2=pypi_0
importlib-metadata=3.10.0=py39hecd8cb5_0
importlib_metadata=3.10.0=hd3eb1b0_0
ipykernel=5.3.4=py39h01d92e1_0
ipython=7.22.0=py39h01d92e1_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
jasper=1.900.1=h636a363_1006
jbig=2.1=h0d85af4_2003
jedi=0.17.2=py39hecd8cb5_1
jinja2=3.0.1=pyhd3eb1b0_0
joblib=1.0.1=pypi_0
jpeg=9d=hbcb3906_0
jsonschema=3.2.0=py_2
jupyter_client=6.1.12=pyhd3eb1b0_0
jupyter_core=4.7.1=py39hecd8cb5_0
jupyterlab_pygments=0.1.2=py_0
keras=2.4.3=pypi_0
keras-nightly=2.5.0.dev2021032900=pypi_0
keras-preprocessing=1.1.2=pypi_0
krb5=1.19.1=hcfbf3a7_0
lame=3.100=h35c211d_1001
lerc=2.2.1=h046ec9c_0
libblas=3.9.0=9_openblas
libcblas=3.9.0=9_openblas
libcurl=7.77.0=hf45b732_0
libcxx=12.0.1=habf9029_0
libdeflate=1.7=h35c211d_5
libedit=3.1.20191231=h0678c8f_2
libev=4.33=haf1e3a3_1
libffi=3.3=hb1e8313_2
libgfortran=5.0.0=9_3_0_h6c81a4c_22
libgfortran5=9.3.0=h6c81a4c_22
libglib=2.68.3=hd556434_0
libiconv=1.16=haf1e3a3_0
liblapack=3.9.0=9_openblas
liblapacke=3.9.0=9_openblas
libnghttp2=1.43.0=h07e645a_0
libopenblas=0.3.15=openmp_h5e1b9a4_1
libopencv=4.5.2=py39h852ad08_1
libpng=1.6.37=h7cec526_2
libprotobuf=3.16.0=hcf210ce_0
libsodium=1.0.18=h1de35cc_0
libssh2=1.9.0=h52ee1ee_6
libtiff=4.3.0=h1167814_1
libwebp-base=1.2.0=h0d85af4_2
libxml2=2.9.12=h93ec3fd_0
llvm-openmp=12.0.1=hda6cdc1_0
lz4-c=1.9.3=h046ec9c_0
markdown=3.3.4=pypi_0
markupsafe=2.0.1=py39h9ed2024_0
mistune=0.8.4=py39h9ed2024_1000
nbclient=0.5.3=pyhd3eb1b0_0
nbconvert=6.1.0=py39hecd8cb5_0
nbformat=5.1.3=pyhd3eb1b0_0
ncurses=6.2=h0a44026_1
nest-asyncio=1.5.1=pyhd3eb1b0_0
nettle=3.6=hedd7734_0
notebook=6.4.0=py39hecd8cb5_0
numpy=1.19.5=pypi_0
oauthlib=3.1.1=pypi_0
opencv=4.5.2=py39h6e9494a_1
openh264=2.1.1=hfd3ada9_0
openssl=1.1.1k=h9ed2024_0
opt-einsum=3.3.0=pypi_0
packaging=21.0=pyhd3eb1b0_0
pandocfilters=1.4.3=py39hecd8cb5_1
parso=0.7.0=py_0
pcre=8.45=he49afe7_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.3.1=pypi_0
pip=21.1.3=py39hecd8cb5_0
pixman=0.40.0=hbcb3906_0
playsound=1.2.2=pypi_0
prometheus_client=0.11.0=pyhd3eb1b0_0
prompt-toolkit=3.0.17=pyh06a4308_0
protobuf=3.17.3=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
py-opencv=4.5.2=py39h71a6800_1
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.20=py_2
pygments=2.9.0=pyhd3eb1b0_0
pyparsing=2.4.7=pyhd3eb1b0_0
pyrsistent=0.18.0=py39h9ed2024_0
python=3.9.5=h88f2d9e_3
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.9=2_cp39
pyyaml=5.4.1=pypi_0
pyzmq=20.0.0=py39h23ab428_1
readline=8.1=h9ed2024_0
requests=2.26.0=pypi_0
requests-oauthlib=1.3.0=pypi_0
rsa=4.7.2=pypi_0
scikit-learn=0.24.2=pypi_0
scipy=1.7.0=pypi_0
send2trash=1.5.0=pyhd3eb1b0_1
setuptools=52.0.0=py39hecd8cb5_0
six=1.15.0=pypi_0
sklearn=0.0=pypi_0
sqlite=3.36.0=hce871da_0
tensorboard=2.5.0=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.0=pypi_0
tensorflow=2.5.0=pypi_0
tensorflow-estimator=2.5.0=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.9.4=py39hecd8cb5_0
testpath=0.5.0=pyhd3eb1b0_0
threadpoolctl=2.2.0=pypi_0
tk=8.6.10=hb0a8c7a_0
tornado=6.1=py39h9ed2024_0
traitlets=5.0.5=pyhd3eb1b0_0
typing-extensions=3.7.4.3=pypi_0
tzdata=2021a=h52ac0ba_0
urllib3=1.26.6=pypi_0
wcwidth=0.2.5=py_0
webencodings=0.5.1=py39hecd8cb5_1
werkzeug=2.0.1=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
wrapt=1.12.1=pypi_0
x264=1!161.3030=h0d85af4_1
xz=5.2.5=h1de35cc_0
zeromq=4.3.4=h23ab428_0
zipp=3.5.0=pyhd3eb1b0_0
zlib=1.2.11=h1de35cc_3
zstd=1.5.0=h582d3a0_0
